+++ src/ec/google/chromeec/smihandler.c
@@ -57,7 +57,7 @@ static void clear_pending_events(void)
 		;
 }
 
-void chromeec_smi_sleep(int slp_type, uint32_t s3_mask, uint32_t s5_mask)
+void chromeec_smi_sleep(int slp_type, uint64_t s3_mask, uint64_t s5_mask)
 {
 	switch (slp_type) {
 	case ACPI_S3:
@@ -78,8 +78,8 @@ void chromeec_smi_sleep(int slp_type, uint32_t s3_mask, uint32_t s5_mask)
 	clear_pending_events();
 }
 
-void chromeec_smi_device_event_sleep(int slp_type, uint32_t s3_mask,
-				     uint32_t s5_mask)
+void chromeec_smi_device_event_sleep(int slp_type, uint64_t s3_mask,
+				     uint64_t s5_mask)
 {
 	switch (slp_type) {
 	case ACPI_S3:
@@ -96,7 +96,7 @@ void chromeec_smi_device_event_sleep(int slp_type, uint32_t s3_mask,
 	google_chromeec_get_device_current_events();
 }
 
-void chromeec_smi_apmc(int apmc, uint32_t sci_mask, uint32_t smi_mask)
+void chromeec_smi_apmc(int apmc, uint64_t sci_mask, uint64_t smi_mask)
 {
 	switch (apmc) {
 	case APM_CNT_ACPI_ENABLE:
